+++ app.py
@@ -1,11 +1,11 @@
 from flask import Flask
 from flask_graphql import GraphQLView
 from commands import init_cli
-from graphene_boilerplate.db import db
+from fga.db import db
 from flask_script import Manager
 from flask_migrate import Migrate
 from flask_graphql_auth import GraphQLAuth
-from graphene_boilerplate.schema import schema
+from fga.schema import schema
 
 
 def create_app():
