+++ fga/gql/auth/__init__.py
@@ -1,0 +1,7 @@
+import graphene
+from fga.gql.auth.mutations import LoginMutation, SignUpMutation
+
+
+class AuthMutation(graphene.ObjectType):
+    login = LoginMutation.Field()
+    sign_up = SignUpMutation.Field()
