+++ fga/model/__init__.py
@@ -1,0 +1,2 @@
+from fga.model.user import User  # noqa: F401
+from fga.model.item import Item  # noqa: F401
