+++ fga/schema.py
@@ -1,0 +1,5 @@
+import graphene
+from fga.gql import MasterQuery
+from fga.gql import MasterMutation
+
+schema = graphene.Schema(query=MasterQuery, mutation=MasterMutation)
