--- graphene_boilerplate/__init__.py
@@ -1,2 +1,0 @@
-from graphene_boilerplate.model import User
-from graphene_boilerplate.model import Item
