--- graphene_boilerplate/model/__init__.py
@@ -1,2 +1,0 @@
-from graphene_boilerplate.model.user import User
-from graphene_boilerplate.model.item import Item
