+++ tests/conftest.py
@@ -5,8 +5,8 @@ from graphene.test import Client as GrapheneClient
 from urllib.parse import urlparse
 
 from app import create_app
-from graphene_boilerplate.db import db
-from graphene_boilerplate.gql import schema
+from fga.db import db
+from fga.gql import schema
 
 
 json_headers = {"Content-Type": "application/json"}
